push

+++ src/guis/ButtonComponent.java
@@ -18,22 +18,23 @@ public class ButtonComponent extends QuadComponent {
 	private Delay delay = new Delay(200);
 	private Text text;
 	private Vector2f textPosition = new Vector2f();
-	
+
 	public ButtonComponent(GUI container, float x, float y, float width, float height) {
 		super(container, x, y, width, height);
 	}
-	
+
 	@Override
 	public void render() {
 		if (Mouse.buttonPressed(button) && clickEvent != null) {
-			Vector2f mouse = new Vector2f(Mouse.getMouseX(), Mouse.getMouseY());
-			if (Collision.rectanglePointCollision(mouse, new Vector4f(getX(), getY(), getWidth(), getHeight())) && delay.activate()) {
+			Vector2f mouse = new Vector2f((float) Mouse.getMouseX(), (float) Mouse.getMouseY());
+			if (Collision.rectanglePointCollision(mouse, new Vector4f(getX(), getY(), getWidth(), getHeight()))
+					&& delay.activate()) {
 				clickEvent.run();
 			}
 		}
-		
+
 		super.render();
-		
+
 		if (text != null) {
 			text.setPosition(Vector.add(new Vector2f(getX(), getY()), textPosition, null));
 			TextMaster.getRenderer().renderText(text);
@@ -55,27 +56,28 @@ public class ButtonComponent extends QuadComponent {
 	public void setClickEvent(Runnable clickEvent) {
 		this.clickEvent = clickEvent;
 	}
-	
+
 	public void setClickDelay(long delay) {
 		this.delay.setDelay(delay);
 	}
-	
+
 	public void setText(String message, String font, float fontSize) {
 		if (text != null)
 			text.delete();
 		text = new Text(message, fontSize, font, new Vector2f(super.getX(), super.getY()), super.getWidth(), true);
 	}
-	
+
 	public void setTextPosition(Vector2f position) {
 		this.textPosition = position;
 	}
-	 
-	@Override public void delete() {
+
+	@Override
+	public void delete() {
 		if (text != null) {
 			text.delete();
 		}
 	}
-	
+
 	public Text getText() {
 		return text;
 	}
