+++ src/guis/TextFieldComponent.java
@@ -14,9 +14,9 @@ import utils.Switch;
 import utils.Vector;
 
 public class TextFieldComponent extends QuadComponent {
-	
+
 	private static TextFieldComponent active;
-	
+
 	private Text text;
 	private String textString = "";
 	private int button = GLFW.GLFW_MOUSE_BUTTON_1;
@@ -24,13 +24,13 @@ public class TextFieldComponent extends QuadComponent {
 	private Text curser;
 	private Switch showCurser = new Switch(500L);
 	private Runnable changeEvent;
-	
+
 	private KeyboardListener listener = new KeyboardListener() {
 		@Override
 		public void keyInput(char c) {
 			if (c == '\b') {
 				if (textString.length() > 0)
-					textString = textString.substring(0, textString.length()-1);
+					textString = textString.substring(0, textString.length() - 1);
 			} else {
 				textString = textString.concat(((Character) c).toString());
 			}
@@ -40,38 +40,40 @@ public class TextFieldComponent extends QuadComponent {
 			text.applyChanges();
 		}
 	};
-	
-	public TextFieldComponent(GUI container, float x, float y, float width, float height, String font, float fontSize, boolean textCentered) {
+
+	public TextFieldComponent(GUI container, float x, float y, float width, float height, String font, float fontSize,
+			boolean textCentered) {
 		super(container, x, y, width, height);
 		text = new Text("", fontSize, font, new Vector2f(x, y), width, textCentered);
 		curser = new Text("|", fontSize, font, new Vector2f(x, y), width, false);
 	}
-	
+
 	public void setTextPosition(Vector2f position) {
 		this.textPosition = position;
 	}
-	
+
 	public void setText(String text) {
 		this.textString = text;
 		this.text.setText(text);
 		this.text.applyChanges();
 	}
-	
-	@Override public void render() {
-		
+
+	@Override
+	public void render() {
+
 		if (!isActive() && Mouse.buttonPressed(button)) {
-			Vector2f mouse = new Vector2f(Mouse.getMouseX(), Mouse.getMouseY());
+			Vector2f mouse = new Vector2f((float) Mouse.getMouseX(), (float) Mouse.getMouseY());
 			if (Collision.rectanglePointCollision(mouse, new Vector4f(getX(), getY(), getWidth(), getHeight()))) {
 				activate();
 			}
 		}
-		
+
 		super.render();
-		
+
 		Vector2f position = Vector.add(new Vector2f(getX(), getY()), textPosition, null);
 		text.setPosition(position);
 		TextMaster.getRenderer().renderText(text);
-		
+
 		if (isActive()) {
 			showCurser.switchBoolean();
 			if (showCurser.getBooleanValue()) {
@@ -80,27 +82,28 @@ public class TextFieldComponent extends QuadComponent {
 			}
 		}
 	}
-	
+
 	public boolean isActive() {
 		return active == this;
 	}
-	
+
 	public void activate() {
 		showCurser.setValue(false);
 		active = this;
 		Keyboard.setListener(listener);
 	}
-	
+
 	public void deactivate() {
 		active = null;
 		Keyboard.removeListener();
 	}
-	
+
 	public String getText() {
 		return text.toString();
 	}
-	
-	@Override public void delete() {
+
+	@Override
+	public void delete() {
 		text.delete();
 		curser.delete();
 	}
